Implement analysis deletion through a command
