+++ tests/Commands/Analysis/DeleteAnalysisCommandTest.php
@@ -1,0 +1,43 @@
+<?php
+
+/*
+ * This file is part of StyleCI.
+ *
+ * (c) Alt Three Services Limited
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace StyleCI\Tests\StyleCI\Commands\Analysis;
+
+use StyleCI\StyleCI\Commands\Analysis\DeleteAnalysisCommand;
+use StyleCI\StyleCI\Handlers\Commands\Analysis\DeleteAnalysisCommandHandler;
+use StyleCI\StyleCI\Models\Analysis;
+use StyleCI\Tests\StyleCI\Commands\AbstractCommandTestCase;
+
+/**
+ * This is the delete analysis command test class.
+ *
+ * @author Graham Campbell <graham@alt-three.com>
+ */
+class DeleteAnalysisCommandTest extends AbstractCommandTestCase
+{
+    protected function getObjectAndParams()
+    {
+        $params = ['analysis' => new Analysis()];
+        $object = new DeleteAnalysisCommand($params['analysis']);
+
+        return compact('params', 'object');
+    }
+
+    protected function objectHasRules()
+    {
+        return false;
+    }
+
+    protected function getHandlerClass()
+    {
+        return DeleteAnalysisCommandHandler::class;
+    }
+}
